requirements.txt
asyncpg
fastapi[all]
httpx
alembic
psycopg2-binary
python-dotenv
passlib
